# Implement byte length calculation that doesn't serialize the bson object